+++ apps/demo-manager/src/main.ts
@@ -2,6 +2,7 @@ import { createApp } from "vue";
 import App from "./App.vue";
 import VirtoShellUi from "@virtoshell/ui";
 import VirtoShellCore from "@virtoshell/core";
+import ModAssets from "@virtoshell/mod-assets";
 import ModOrders from "./modules/orders";
 import ModProducts from "./modules/products";
 import ModOffers from "./modules/offers";
@@ -14,6 +15,7 @@ import "@virtoshell/ui/dist/ui.css";
 const app = createApp(App)
   .use(VirtoShellUi)
   .use(VirtoShellCore)
+  .use(ModAssets)
   .use(ModOrders)
   .use(ModProducts)
   .use(ModOffers);