+++ apps/demo-manager/src/modules/orders/composables/useOrders/index.ts
@@ -5,6 +5,7 @@ import {
   OrderModuleClient,
   CustomerOrderSearchResult,
   CustomerOrderSearchCriteria,
+  ICustomerOrderSearchCriteria,
 } from "@virtoshell/api-client";
 
 const orders: Ref<CustomerOrderSearchResult> = ref(
@@ -12,18 +13,28 @@ const orders: Ref<CustomerOrderSearchResult> = ref(
 );
 
 export default () => {
-  async function loadOrders() {
-    var { getAccessToken } = useUser();
+  const loading = ref(false);
+  const currentPage = ref(1);
+
+  async function loadOrders(query?: ICustomerOrderSearchCriteria) {
+    loading.value = true;
+    const { getAccessToken } = useUser();
     const client = new OrderModuleClient();
 
     client.setAuthToken(await getAccessToken());
     orders.value = await client.searchCustomerOrder({
       take: 20,
+      ...(query || {})
     } as CustomerOrderSearchCriteria);
+    currentPage.value = ((query?.skip || 0) / Math.max(1, query?.take || 20)) + 1
+    loading.value = false;
   }
 
   return {
     orders: computed(() => orders.value),
+    pages: computed(() => Math.ceil(orders.value?.totalCount / 20)),
+    currentPage,
+    loading,
     loadOrders,
   };
 };